Split assignments tests to improve parallel testing performance
