+++ test/unit/assignments/assignments.spec.ts
@@ -1,0 +1,181 @@
+import * as ts from "typescript";
+import { TSTLErrors } from "../../../src/TSTLErrors";
+import * as util from "../../util";
+
+test.each([
+    { inp: `"abc"`, out: `"abc"` },
+    { inp: "3", out: "3" },
+    { inp: "[1,2,3]", out: "{1, 2, 3}" },
+    { inp: "true", out: "true" },
+    { inp: "false", out: "false" },
+    { inp: `{a:3,b:"4"}`, out: `{a = 3, b = "4"}` },
+])("Const assignment (%p)", ({ inp, out }) => {
+    const lua = util.transpileString(`const myvar = ${inp};`);
+    expect(lua).toBe(`local myvar = ${out};`);
+});
+
+test.each([
+    { inp: `"abc"`, out: `"abc"` },
+    { inp: "3", out: "3" },
+    { inp: "[1,2,3]", out: "{1, 2, 3}" },
+    { inp: "true", out: "true" },
+    { inp: "false", out: "false" },
+    { inp: `{a:3,b:"4"}`, out: `{a = 3, b = "4"}` },
+])("Let assignment (%p)", ({ inp, out }) => {
+    const lua = util.transpileString(`let myvar = ${inp};`);
+    expect(lua).toBe(`local myvar = ${out};`);
+});
+
+test.each([
+    { inp: `"abc"`, out: `"abc"` },
+    { inp: "3", out: "3" },
+    { inp: "[1,2,3]", out: "{1, 2, 3}" },
+    { inp: "true", out: "true" },
+    { inp: "false", out: "false" },
+    { inp: `{a:3,b:"4"}`, out: `{a = 3, b = "4"}` },
+])("Var assignment (%p)", ({ inp, out }) => {
+    const lua = util.transpileString(`var myvar = ${inp};`);
+    expect(lua).toBe(`myvar = ${out};`);
+});
+
+test.each(["var myvar;", "let myvar;", "const myvar = null;", "const myvar = undefined;"])(
+    "Null assignments (%p)",
+    declaration => {
+        const result = util.transpileAndExecute(declaration + " return myvar;");
+        expect(result).toBe(undefined);
+    },
+);
+
+test.each([
+    { input: ["a", "b"], values: ["e", "f"] },
+    { input: ["a", "b"], values: ["e", "f", "g"] },
+    { input: ["a", "b", "c"], values: ["e", "f", "g"] },
+])("Binding pattern assignment (%p)", ({ input, values }) => {
+    const pattern = input.join(",");
+    const initializer = values.map(v => `"${v}"`).join(",");
+
+    const tsCode = `const [${pattern}] = [${initializer}]; return [${pattern}].join("-");`;
+    const result = util.transpileAndExecute(tsCode);
+
+    expect(result).toBe(values.slice(0, input.length).join("-"));
+});
+
+test("Ellipsis binding pattern", () => {
+    expect(() => util.transpileString("let [a,b,...c] = [1,2,3];")).toThrowExactError(
+        TSTLErrors.ForbiddenEllipsisDestruction(util.nodeStub),
+    );
+});
+
+test("Tuple Assignment", () => {
+    const code = `function abc(): [number, number] { return [1, 2]; };
+                  let t: [number, number] = abc();
+                  return t[0] + t[1];`;
+    const result = util.transpileAndExecute(code);
+    expect(result).toBe(3);
+});
+
+test("TupleReturn assignment", () => {
+    const code =
+        `/** @tupleReturn */\n` +
+        `declare function abc(this: void): number[]\n` +
+        `let [a,b] = abc();`;
+
+    const lua = util.transpileString(code);
+    expect(lua).toBe("local a, b = abc();");
+});
+
+test("TupleReturn Single assignment", () => {
+    const code =
+        `/** @tupleReturn */\n` +
+        `declare function abc(this: void): [number, string];\n` +
+        `let a = abc();` +
+        `a = abc();`;
+
+    const lua = util.transpileString(code);
+    expect(lua).toBe("local a = ({abc()});\na = ({abc()});");
+});
+
+test("TupleReturn interface assignment", () => {
+    const code =
+        `interface def {\n` +
+        `/** @tupleReturn */\n` +
+        `abc();\n` +
+        `} declare const jkl : def;\n` +
+        `let [a,b] = jkl.abc();`;
+
+    const lua = util.transpileString(code);
+    expect(lua).toBe("local a, b = jkl:abc();");
+});
+
+test("TupleReturn namespace assignment", () => {
+    const code =
+        `declare namespace def {\n` +
+        `/** @tupleReturn */\n` +
+        `function abc(this: void) {}\n` +
+        `}\n` +
+        `let [a,b] = def.abc();`;
+
+    const lua = util.transpileString(code);
+    expect(lua).toBe("local a, b = def.abc();");
+});
+
+test("TupleReturn method assignment", () => {
+    const code =
+        `declare class def {\n` +
+        `/** @tupleReturn */\n` +
+        `abc() { return [1,2,3]; }\n` +
+        `} const jkl = new def();\n` +
+        `let [a,b] = jkl.abc();`;
+
+    const lua = util.transpileString(code);
+    expect(lua).toBe("local jkl = def.new();\nlocal a, b = jkl:abc();");
+});
+
+test("TupleReturn functional", () => {
+    const code = `/** @tupleReturn */
+    function abc(): [number, string] { return [3, "a"]; }
+    const [a, b] = abc();
+    return b + a;`;
+
+    const result = util.transpileAndExecute(code);
+
+    expect(result).toBe("a3");
+});
+
+test("TupleReturn single", () => {
+    const code = `/** @tupleReturn */
+    function abc(): [number, string] { return [3, "a"]; }
+    const res = abc();
+    return res.length`;
+
+    const result = util.transpileAndExecute(code);
+
+    expect(result).toBe(2);
+});
+
+test("TupleReturn in expression", () => {
+    const code = `/** @tupleReturn */
+    function abc(): [number, string] { return [3, "a"]; }
+    return abc()[1] + abc()[0];`;
+
+    const result = util.transpileAndExecute(code);
+
+    expect(result).toBe("a3");
+});
+
+test.each(["and", "local", "nil", "not", "or", "repeat", "then", "until"])(
+    "Keyword identifier error (%p)",
+    identifier => {
+        expect(() => util.transpileString(`const ${identifier} = 3;`)).toThrowExactError(
+            TSTLErrors.KeywordIdentifier(ts.createIdentifier(identifier)),
+        );
+    },
+);
+
+test("String table access", () => {
+    const code = `const dict : {[key:string]:any} = {};
+                  dict["a b"] = 3;
+                  return dict["a b"];`;
+    const result = util.transpileAndExecute(code);
+    expect(result).toBe(3);
+});
